Add files via upload
This is the updated database. I have created the comparison table using a foreign key within the database. I still need to insert the values from the database for the profile.
and ect. However, this functionality will be present after our other Engineers commit code for the interface. There might still be more to be added in terms of tables and other comparison table however this will be evaluated after further team collaboration.