Add files via upload
This is the Java code we will be using to create a connection between our interface and our database. This has been created using a online tutorial. From the knowledge gained so far as well as the code, we will have an object which will be our connection. Lets call it a plug and the actual connection function takes our host, ip , password and ect and within that use it to accces our database with a connection function